pyproject.toml
[project]
name = "sprintsync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "boto3>=1.40.20",
    "cachetools>=6.2.0",
    "cryptography>=45.0.6",
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "structlog>=24.4.0",
    "python-json-logger>=2.0.7",
]
